Main class cleanup and added SpawnService
