+++ plugin/src/main/java/pl/mgtm/magicznakraina/MagicznaKraina.java
@@ -1,18 +1,15 @@
 package pl.mgtm.magicznakraina;
 
 import org.bukkit.ChatColor;
-import org.bukkit.Location;
-import org.bukkit.World;
 import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
 import pl.mgtm.magicznakraina.commands.*;
 import pl.mgtm.magicznakraina.events.DeathEvent;
 import pl.mgtm.magicznakraina.events.JoinServerEvent;
 import pl.mgtm.magicznakraina.events.RespawnEvent;
 import pl.mgtm.magicznakraina.modules.protect_chests.ProtectChests;
+import pl.mgtm.magicznakraina.services.SpawnService;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 
 public final class MagicznaKraina extends JavaPlugin {
@@ -21,37 +18,23 @@ public final class MagicznaKraina extends JavaPlugin {
 
     private FileConfiguration config;
 
-    // Przechowuje wszystkich graczy, do ktorych chcemy sie przeteleportowac
-    // Player = gracz do ktorego chca sie przeteleportowac np. kris
-    // ArrayList<Player> = lista graczy, ktorzy wyslali prosbe o teleportacje
-    private HashMap<Player, ArrayList<Player>> tpa = new HashMap<Player, ArrayList<Player>>();
-
-    // Lokalizacja spawnu serwera
-    private Location spawnLocation;
+    public SpawnService spawnService;
 
     private HashMap<String, String> messages = new HashMap<>();
 
-    // Wyswietla pomocnicze informacje
-    // Przydatne podczas tworzenia pluginow
+    // TODO: Instrumenty i profiler
     private boolean debug = false;
 
 
     @Override
     public void onEnable() {
-        // ===================================================== //
-        //                                                       //
-        //                       Plugin info                     //
-        //                                                       //
-        // ===================================================== //
-
+        // Remove brackets and trailing spaces from partially initialized arrays
         String pluginAuthor = getDescription().getAuthors().toString()
-                .replace("[", "")  // remove the right bracket
-                .replace("]", "")  // remove the left bracket
-                .trim();           // remove trailing spaces from partially initialized arrays
+                .replace("[", "")
+                .replace("]", "")
+                .trim();
 
-        // Plugin startup logic
-        getLogger().info(ChatColor.GREEN + "Plugin " + getName() + " v" + getDescription().getVersion() + " loading...");
-        getLogger().info("Author: " + pluginAuthor);
+        getLogger().info("Loading MagicznaKraina (v" + getDescription().getVersion() + " - " + pluginAuthor + ")");
 
         setInstance(this);
 
@@ -80,45 +63,27 @@ public final class MagicznaKraina extends JavaPlugin {
         //                                                       //
         // ===================================================== //
 
-        // ProtectChests
-        new ProtectChests();
+        this.spawnService = new SpawnService();
 
-        // Zaladuj config
+        new ProtectChests();
         loadConfig();
-        // Zaladuj spawn
-        loadSpawn();
+
+        this.spawnService.loadSpawnLocation();
 
         messages.put("leaveMessage", ChatColor.translateAlternateColorCodes('&', getConfig().getString("message.leaveMessage")));
         messages.put("joinMessage", ChatColor.translateAlternateColorCodes('&', getConfig().getString("message.joinMessage")));
 
-        // Get debug informations
         debug = getConfig().getBoolean("debug");
 
-        getLogger().info(ChatColor.GREEN + "Plugin loaded!");
-
-
-
+        getLogger().info("MagicznaKraina has been successfully loaded!");
     }
 
     @Override
     public void onDisable() {
-        // Plugin shutdown logic
-        getLogger().info(ChatColor.GREEN + "Plugin shutdown.");
+        getLogger().info("MagicznaKraina has been shutdown!");
     }
 
-    private void loadSpawn() {
-        if (config.contains("spawn.x")) {
-            String worldName = config.getString("spawn.world");
-            double x = config.getDouble("spawn.x");
-            double y = config.getDouble("spawn.y");
-            double z = config.getDouble("spawn.z");
-            float pitch = (float) config.getDouble("spawn.pitch");
-            float yaw = (float) config.getDouble("spawn.yaw");
-
-            spawnLocation = new Location(getServer().getWorld(worldName), x, y, z, yaw, pitch);
-        }
-    }
-
+    // TODO: Move it somewhere else
     private void loadConfig() {
         getConfig().options().copyDefaults(true);
         saveConfig();
@@ -136,26 +101,12 @@ public final class MagicznaKraina extends JavaPlugin {
         MagicznaKraina.instance = instance;
     }
 
-    public HashMap<Player, ArrayList<Player>> getTpa() {
-        return tpa;
-    }
-
+    // TODO: Move it somewhere else
     public HashMap<String, String> getMessages() {
         return messages;
     }
 
-    public Location getSpawnLocation() {
-        return spawnLocation;
-    }
-
-    public void setSpawnLocation(Location spawnLocation) {
-        this.spawnLocation = spawnLocation;
-    }
-
-    public void setSpawnLocation(World world, double x, double y, double z, float yaw, float pitch) {
-        this.spawnLocation = new Location(world, x, y, z, yaw, pitch);
-    }
-
+    // TODO: ... nothing to explain in here
     public void setDebug(boolean debug) {
         this.debug = debug;
     }
